test(mongo): clean up tests for mongo provider
